+++ adapters/BidMachineAdapter/CHANGELOG.md
@@ -1,1 +1,4 @@
 ## BidMachine iOS Mediation Adapter Changelog
+
+#### Next version
+- Created BidMachineAdapter that support bidding banner, interstitial, rewarded, and native ads.